Create PhyloPhlAn.sh
Give a genome and calculate the genomic distance between the genome and the SGBs in the database. If the distance is <5% the genome can be considered part of the SGB. If a genome is >5% distance to all the sequences you have found a new organism not present in the database.